change return value of TotalProcesses()

+++ include/system.h
@@ -14,7 +14,7 @@ public:
   std::vector<Process>& Processes();
   float MemoryUtilization();
   long UpTime();
-  int TotalProcesses(); 
+  long TotalProcesses(); 
   int RunningProcesses();
   std::string Kernel();
   std::string OperatingSystem();
